requirements.txt
langchain>=0.3.0
langchain-mistralai>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.5.0
duckduckgo-search>=6.3.0
pydantic>=2.9.0
python-dotenv>=1.0.1
twilio>=9.3.0
jinja2>=3.1.4
